20200331
1. 충돌 처리 추가
양 옆 벽에 부딪히면 입력시 좌우 이동 불가, 맨 아래 도착시 보드에 블록 배치 확정

+++ Project_Beom/Project_Beom/GameObject.cpp
@@ -20,12 +20,12 @@ void GameObject::SetPos(int x, int y)
 
 void GameObject::SetPosX(int x)
 {
-	m_posX = (int)x;
+	m_posX = (float)x;
 }
 
 void GameObject::SetPosY(int y)
 {
-	m_posY = (int)y;
+	m_posY = (float)y;
 }
 
 void GameObject::SetSpeed(float speed)
